chore: consolidate worker code
This patch consolidates all the worker code into a single file, as it
was no longer used in multiple places.
